+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 2.1.0
+
+Adds keybindings for going up/down through the set of stories, this only loops in the same section.
+As I can't change the current selection inside the app, I'm not putting these in the README. You can 
+use `cmd + k, up` or `cmd + k, down` to change them. - orta
+
 ### 2.0.0
 
 Adds storybooks inline, so you don't need to use the site at all now. Tab still exists, but useless. - orta